Add documentation to SobolEngine, rename file as quasirandom.py
